# Conflicts:
#	src/EPPlus/Drawing/Chart/ExcelChartSeries.cs
#	src/EPPlus/Drawing/OleObject/ExcelOleObject.cs
#	src/EPPlus/FormulaParsing/DependencyChain/RpnFormulaExecution.cs
#	src/EPPlus/FormulaParsing/Excel/Functions/MathFunctions/AverageIf.cs
#	src/EPPlus/FormulaParsing/Excel/Functions/MathFunctions/AverageIfs.cs
#	src/EPPlus/FormulaParsing/Excel/Functions/MathFunctions/SumIfs.cs
#	src/EPPlus/FormulaParsing/ExcelUtilities/RangeFlattener.cs
#	src/EPPlus/Utils/TypeConversion/ConvertUtil.cs
#	src/EPPlus/Utils/XML/XmlReaderHelper.cs
#	src/EPPlusTest/Drawing/PictureTests.cs

+++ src/EPPlus/Table/ExcelTableDxfBase.cs
@@ -287,7 +287,7 @@ namespace OfficeOpenXml.Table
                     color.SetColor((ExcelIndexedColor)value);
                     break;
                 case eStyleProperty.Tint:
-                    color.Tint = (decimal)value;
+                    color.Tint = (double)value;
                     break;
             }
         }
